Fix bugs and improve spawning locations for pursuers/evader
Fixed bugs with not ending the game properly and with spawning
at edge of board (where there's a wall). Also, improved spawning
locations for pursuers / evaders by not allowing them to spawn at
the same location, by enforcing that pursuers can actually reach
the evader, and that evaders spawn in locations with at least
two possible moves (to avoid the evader doing nothing).